requirements.txt
llama-index
llama-index-embeddings-huggingface
llama-index-llms-ollama
llama-index-llms-groq
llama-index-vector-stores-qdrant
llama-index-vector-stores-chroma
qdrant-client
chromadb
sentence-transformers
ragas
streamlit
fastapi
uvicorn
python-dotenv
pypdf
jupyterlab
langchain-groq
langchain-huggingface
